requirements/base.txt
Django==1.7.1
Pillow==2.5.0
PyJWT==0.3.0
PyPDF2==1.23
app-version==0.2.1
appdirs==1.4.3
argparse==1.2.1
asn1crypto==0.22.0
certifi==2017.7.27.1
chardet==3.0.4
cryptography==2.1.4
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
django-braces==1.4.0
django-cuser==2014.9.28
django-dirtyfields==0.3
git+https://github.com/jbroda/django-dropbox.git@master
django-haystack==2.3.1
###django-inspectional-registration==0.4.5
django-inspectional-registration==0.6.2
django-localflavor==1.0
django-registration-redux==1.1
django-reversion==1.8.5
django-reversion-compare==0.5.1
django-sslserver==0.14
docutils==0.12
dropbox==8.0.0
elasticsearch==1.2.0
enum34==1.1.6
gdata==2.0.18
html5lib==0.999
httplib2==0.9
idna==2.6
ipaddress==1.0.18
packaging==16.8
pbr==0.10.0
psycopg2==2.7.4
pyelasticsearch==0.7.1
pyopenssl==17.5.0
pyparsing==2.2.0
pystache==0.5.4
python-dateutil==2.2
pytz==2014.9
reportlab==2.7
requests==2.18.4
Scrapy==1.5.0
simplejson==3.6.5
six==1.11.0
sqlparse==0.1.13
static==1.1.1
static3==0.5.1
stevedore==1.1.0
tolerance==0.1.2
typing==3.5.3.0
urllib3==1.22
wsgiref==0.1.2
xhtml2pdf==0.0.6
